% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_string.R
\name{reverse_string}
\alias{reverse_string}
\title{Reverse a string}
\usage{
reverse_string(str)
}
\arguments{
\item{str}{A single string to be reversed.}
}
\value{
The reversed string.
}
\description{
This function takes a single string as an argument and returns its reverse.
}
\examples{
reverse_string("hello")
}
