% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_case.R
\name{to_uppercase}
\alias{to_uppercase}
\title{Convert a string to uppercase}
\usage{
to_uppercase(str)
}
\arguments{
\item{str}{A single string to be converted to uppercase.}
}
\value{
The uppercase string.
}
\description{
This function takes a single string as an argument and returns the string in uppercase.
}
\examples{
to_uppercase("hello")
}
